requirements.txt
numpy
pandas
openpyxl
spacy
geonamescache
rapidfuzz
boto3
python-dotenv
